Update README.md (#10731)
* Update README.md

I updated a few links that pointed to outdated content and updated quite a few links that were hitting redirects -- I changed them to the correct destination.  I'd also like to update the primary visual to an updated version.  I'm not sure how to do that.  We can use the most recent 'press' screenshots that the PM team (Jason) developed.

* Update README.md