MM-16378: revert plugin health check config removal (#11819)
* revert 4e5f6fcfbc85ea24788241c716862e1833c6f60e

This reverts the configuration setting to disable plugin health checks, preferring instead to retain this functionality for advanced cases. In part, this was driven by the discovery that the health checks were failing on Windows, though that will be addressed separately.

Fixes: MM-16378

* simplify interval handling/logging

Also shutdown the job when plugins are altogether disabled.